+++ src/core/chuck_lang.h
@@ -59,15 +59,6 @@ t_CKBOOL init_class_array( Chuck_Env * env, Chuck_Type * type );
 t_CKBOOL init_class_vec3( Chuck_Env * env, Chuck_Type * type ); // 1.3.5.3
 t_CKBOOL init_class_vec4( Chuck_Env * env, Chuck_Type * type ); // 1.3.5.3
 
-t_CKBOOL init_class_Midi( Chuck_Env * env );
-t_CKBOOL init_class_MidiRW( Chuck_Env * env );
-t_CKBOOL init_class_HID( Chuck_Env * env );
-t_CKBOOL init_class_io( Chuck_Env * env, Chuck_Type * type );
-t_CKBOOL init_class_fileio( Chuck_Env * env, Chuck_Type * type );
-// added 1.3.0.0 -- moved to be full-fledged class
-t_CKBOOL init_class_chout( Chuck_Env * env, Chuck_Type * type );
-// added 1.3.0.0 -- moved to be full-fledged class
-t_CKBOOL init_class_cherr( Chuck_Env * env, Chuck_Type * type );
 
 
 
@@ -205,84 +196,6 @@ CK_DLL_MFUN( event_can_wait );
 
 
 //-----------------------------------------------------------------------------
-// io API
-//-----------------------------------------------------------------------------
-CK_DLL_MFUN( io_dummy );
-CK_DLL_SFUN( io_newline );
-CK_DLL_SFUN( io_openfile );
-
-
-//-----------------------------------------------------------------------------
-// fileio API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( fileio_ctor );
-CK_DLL_DTOR( fileio_dtor );
-CK_DLL_MFUN( fileio_open );
-CK_DLL_MFUN( fileio_openflags );
-CK_DLL_MFUN( fileio_good );
-CK_DLL_MFUN( fileio_close );
-CK_DLL_MFUN( fileio_flush );
-CK_DLL_MFUN( fileio_getmode );
-CK_DLL_MFUN( fileio_setmode );
-CK_DLL_MFUN( fileio_size );
-CK_DLL_MFUN( fileio_seek );
-CK_DLL_MFUN( fileio_tell );
-CK_DLL_MFUN( fileio_isdir );
-CK_DLL_MFUN( fileio_dirlist );
-CK_DLL_MFUN( fileio_read );
-CK_DLL_MFUN( fileio_readline );
-CK_DLL_MFUN( fileio_readint );
-CK_DLL_MFUN( fileio_readintflags );
-CK_DLL_MFUN( fileio_readfloat );
-CK_DLL_MFUN( fileio_eof );
-CK_DLL_MFUN( fileio_more );
-CK_DLL_MFUN( fileio_writestring );
-CK_DLL_MFUN( fileio_writeint );
-CK_DLL_MFUN( fileio_writeintflags );
-CK_DLL_MFUN( fileio_writefloat );
-
-
-//-----------------------------------------------------------------------------
-// chout API -- added 1.3.0.0 as full class
-//-----------------------------------------------------------------------------
-CK_DLL_MFUN( chout_good );
-CK_DLL_MFUN( chout_close );
-CK_DLL_MFUN( chout_flush );
-CK_DLL_MFUN( chout_getmode );
-CK_DLL_MFUN( chout_setmode );
-CK_DLL_MFUN( chout_read );
-CK_DLL_MFUN( chout_readline );
-CK_DLL_MFUN( chout_readint );
-CK_DLL_MFUN( chout_readintflags );
-CK_DLL_MFUN( chout_readfloat );
-CK_DLL_MFUN( chout_eof );
-CK_DLL_MFUN( chout_more );
-CK_DLL_MFUN( chout_writestring );
-CK_DLL_MFUN( chout_writeint );
-CK_DLL_MFUN( chout_writefloat );
-
-
-//-----------------------------------------------------------------------------
-// cherr API -- added 1.3.0.0 as full class
-//-----------------------------------------------------------------------------
-CK_DLL_MFUN( cherr_good );
-CK_DLL_MFUN( cherr_close );
-CK_DLL_MFUN( cherr_flush );
-CK_DLL_MFUN( cherr_getmode );
-CK_DLL_MFUN( cherr_setmode );
-CK_DLL_MFUN( cherr_read );
-CK_DLL_MFUN( cherr_readline );
-CK_DLL_MFUN( cherr_readint );
-CK_DLL_MFUN( cherr_readintflags );
-CK_DLL_MFUN( cherr_readfloat );
-CK_DLL_MFUN( cherr_eof );
-CK_DLL_MFUN( cherr_more );
-CK_DLL_MFUN( cherr_writestring );
-CK_DLL_MFUN( cherr_writeint );
-CK_DLL_MFUN( cherr_writefloat );
-
-
-//-----------------------------------------------------------------------------
 // string API
 //-----------------------------------------------------------------------------
 CK_DLL_CTOR( string_ctor );
@@ -353,127 +266,6 @@ public:
 // exception API
 //-----------------------------------------------------------------------------
 
-
-//-----------------------------------------------------------------------------
-// MidiMsg API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( MidiMsg_ctor );
-
-extern t_CKUINT MidiMsg_offset_data1;
-extern t_CKUINT MidiMsg_offset_data2;
-extern t_CKUINT MidiMsg_offset_data3;
-extern t_CKUINT MidiMsg_offset_when;
-
-//-----------------------------------------------------------------------------
-// MidiRW API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( MidiRW_ctor );
-CK_DLL_DTOR( MidiRW_dtor );
-CK_DLL_MFUN( MidiRW_open );
-CK_DLL_MFUN( MidiRW_close );
-CK_DLL_MFUN( MidiRW_read );
-CK_DLL_MFUN( MidiRW_write );
-
-
-//-----------------------------------------------------------------------------
-// MidiMsgOut API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( MidiMsgOut_ctor );
-CK_DLL_DTOR( MidiMsgOut_dtor );
-CK_DLL_MFUN( MidiMsgOut_open );
-CK_DLL_MFUN( MidiMsgOut_close );
-CK_DLL_MFUN( MidiMsgOut_write );
-
-
-//-----------------------------------------------------------------------------
-// MidiMsgIn API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( MidiMsgIn_ctor );
-CK_DLL_DTOR( MidiMsgIn_dtor );
-CK_DLL_MFUN( MidiMsgIn_open );
-CK_DLL_MFUN( MidiMsgIn_close );
-CK_DLL_MFUN( MidiMsgIn_read );
-
-
-//-----------------------------------------------------------------------------
-// MidiIn API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( MidiIn_ctor );
-CK_DLL_DTOR( MidiIn_dtor );
-CK_DLL_MFUN( MidiIn_open );
-CK_DLL_MFUN( MidiIn_open_named ); // added 1.3.0.0
-CK_DLL_MFUN( MidiIn_open_named_i ); // added 1.3.0.0
-CK_DLL_MFUN( MidiIn_good );
-CK_DLL_MFUN( MidiIn_num );
-CK_DLL_MFUN( MidiIn_name );
-CK_DLL_MFUN( MidiIn_printerr );
-CK_DLL_MFUN( MidiIn_recv );
-CK_DLL_MFUN( MidiIn_can_wait );
-
-
-//-----------------------------------------------------------------------------
-// MidiOut API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( MidiOut_ctor );
-CK_DLL_DTOR( MidiOut_dtor );
-CK_DLL_MFUN( MidiOut_open );
-CK_DLL_MFUN( MidiOut_open_named ); // added 1.3.0.0
-CK_DLL_MFUN( MidiOut_open_named_i); // added 1.3.0.0
-CK_DLL_MFUN( MidiOut_good );
-CK_DLL_MFUN( MidiOut_num );
-CK_DLL_MFUN( MidiOut_name );
-CK_DLL_MFUN( MidiOut_printerr );
-CK_DLL_MFUN( MidiOut_send );
-
-
-//-----------------------------------------------------------------------------
-// HidMsg API
-//-----------------------------------------------------------------------------
-CK_DLL_MFUN( HidMsg_is_axis_motion );
-CK_DLL_MFUN( HidMsg_is_button_down );
-CK_DLL_MFUN( HidMsg_is_button_up );
-CK_DLL_MFUN( HidMsg_is_mouse_motion );
-CK_DLL_MFUN( HidMsg_is_hat_motion );
-CK_DLL_MFUN( HidMsg_is_wheel_motion );
-
-//-----------------------------------------------------------------------------
-// HidIn API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( HidIn_ctor );
-CK_DLL_DTOR( HidIn_dtor );
-CK_DLL_MFUN( HidIn_open );
-CK_DLL_MFUN( HidIn_open_named );
-CK_DLL_MFUN( HidIn_open_named_i ); // added 1.3.0.0
-CK_DLL_MFUN( HidIn_open_joystick );
-CK_DLL_MFUN( HidIn_open_mouse );
-CK_DLL_MFUN( HidIn_open_keyboard );
-CK_DLL_MFUN( HidIn_open_tiltsensor );
-CK_DLL_MFUN( HidIn_good );
-CK_DLL_MFUN( HidIn_num );
-CK_DLL_MFUN( HidIn_name );
-CK_DLL_MFUN( HidIn_printerr );
-CK_DLL_MFUN( HidIn_recv );
-CK_DLL_MFUN( HidIn_read );
-CK_DLL_MFUN( HidIn_send );
-CK_DLL_MFUN( HidIn_can_wait );
-CK_DLL_SFUN( HidIn_read_tilt_sensor );
-CK_DLL_SFUN( HidIn_ctrl_tiltPollRate );
-CK_DLL_SFUN( HidIn_cget_tiltPollRate );
-CK_DLL_SFUN( HidIn_start_cursor_track );
-CK_DLL_SFUN( HidIn_stop_cursor_track );
-
-
-//-----------------------------------------------------------------------------
-// HidOut API
-//-----------------------------------------------------------------------------
-CK_DLL_CTOR( HidOut_ctor );
-CK_DLL_DTOR( HidOut_dtor );
-CK_DLL_MFUN( HidOut_open );
-CK_DLL_MFUN( HidOut_good );
-CK_DLL_MFUN( HidOut_num );
-CK_DLL_MFUN( HidOut_name );
-CK_DLL_MFUN( HidOut_printerr );
-CK_DLL_MFUN( HidOut_send );
